pyproject.toml
[project]
name = "rag-qdrant"
version = "0.1.0"
description = "End-to-end RAG with Qdrant, Voyage AI, and Claude"
requires-python = ">=3.11"
dependencies = [
    # Core API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",  # required by FastAPI for UploadFile / multipart forms
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    # Vector store
    "qdrant-client>=1.12.0",
    # Embeddings & LLM
    "voyageai>=0.3.0",
    "anthropic>=0.39.0",
    # PDF parsing
    "pypdf>=5.1.0",
    "pdfplumber>=0.11.4",
    # Chunking
    "langchain-text-splitters>=0.3.2",
    # UI
    "streamlit>=1.40.0",
    "httpx>=0.27.2",
    # CLI / utility
    "typer>=0.13.0",
    "rich>=13.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]
# Orchestration extras — only needed when running the Dagster ingestion pipeline.
# Install with: uv sync --group dagster
dagster = [
    "dagster>=1.9.0",
    "dagster-webserver>=1.9.0",
    "azure-storage-blob>=12.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.dagster]
module_name = "rag_pipelines.definitions"
code_location_name = "rag_qdrant_ingest"
